changed population

+++ Data/create_table.sql
@@ -2,7 +2,7 @@ DROP TABLE IF EXISTS water_country;
 CREATE TABLE water_country (
     country TEXT,
     year SMALLINT,
-    population_thousands INTEGER,
+    population_thousands REAL,
     rural_available REAL,
     rural_contam_free REAL,
     urban_available REAL,
@@ -15,7 +15,7 @@ DROP TABLE IF EXISTS water_region;
 CREATE TABLE water_region (
     country TEXT,
     year SMALLINT,
-    population_thousands INTEGER,
+    population_thousands REAL,
     rural_available REAL,
     rural_contam_free REAL,
     urban_available REAL,
